+++ plugin/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.4.3]
+- Added video fit feature by [Utkarsh Sharma](https://github.com/uttusharma). 
+
+## [0.4.2]
+- Fixed loading animation autoplay false.
+
 ## [0.4.1+1]
 - Removed logs.
 ## [0.4.1]